requirements.txt
flask
google-genai
chainlit
httpx
